refactor expandedPopups -> popupActions
